Rename packages and perform some test refactor
