--- app/src/main/java/com/example/vonbass/quotegenerator_mvp/presenter/MainPresenter.java
@@ -1,11 +1,0 @@
-package com.example.vonbass.quotegenerator_mvp.presenter;
-
-/**
- * Created by VonBass on 5/1/2018.
- */
-public interface MainPresenter {
-
-    void onButtonClick();
-
-    void onDestroy();
-}
