--- app/src/main/java/com/example/vonbass/quotegenerator_mvp/view/MainView.java
@@ -1,13 +1,0 @@
-package com.example.vonbass.quotegenerator_mvp.view;
-
-/**
- * Created by VonBass on 5/1/2018.
- */
-public interface MainView {
-
-    void showProgress();
-
-    void hideProgress();
-
-    void setQuote(String string);
-}
