+++ app/src/main/java/org/smartregister/mvp/random_quote_generator/presenter/MainPresenter.java
@@ -1,0 +1,11 @@
+package org.smartregister.mvp.random_quote_generator.presenter;
+
+/**
+ * Created by VonBass on 5/1/2018.
+ */
+public interface MainPresenter {
+
+    void onButtonClick();
+
+    void onDestroy();
+}
