+++ app/src/main/java/org/smartregister/mvp/random_quote_generator/view/MainView.java
@@ -1,0 +1,13 @@
+package org.smartregister.mvp.random_quote_generator.view;
+
+/**
+ * Created by VonBass on 5/1/2018.
+ */
+public interface MainView {
+
+    void showProgress();
+
+    void hideProgress();
+
+    void setQuote(String string);
+}
