+++ app/src/main/java/org/smartregister/mvp/random_quote_generator/view/activity/MainActivity.java
@@ -1,0 +1,69 @@
+package org.smartregister.mvp.random_quote_generator.view.activity;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.ProgressBar;
+import android.widget.TextView;
+
+import org.smartregister.mvp.random_quote_generator.view.MainView;
+import org.smartregister.mvp.random_quote_generator.R;
+import org.smartregister.mvp.random_quote_generator.presenter.MainPresenter;
+import org.smartregister.mvp.random_quote_generator.presenter.MainPresenterImpl;
+
+import static android.view.View.GONE;
+
+public class MainActivity extends AppCompatActivity implements MainView {
+
+    private TextView textView;
+    private Button button;
+    private ProgressBar progressBar;
+    MainPresenter presenter;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        textView = (TextView) findViewById(R.id.textView);
+        button = (Button) findViewById(R.id.button);
+        progressBar = (ProgressBar) findViewById(R.id.progressBar);
+        presenter = new MainPresenterImpl(this);
+
+        button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                presenter.onButtonClick();
+            }
+        });
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        presenter.onDestroy();
+    }
+
+    @Override
+    public void showProgress() {
+        progressBar.setVisibility(View.VISIBLE);
+        textView.setVisibility(View.INVISIBLE);
+    }
+
+    @Override
+    public void hideProgress() {
+        progressBar.setVisibility(GONE);
+        textView.setVisibility(View.VISIBLE);
+    }
+
+    @Override
+    public void setQuote(String string) {
+        textView.setText(string);
+    }
+}
